requirements.txt
evaluate==0.4.6
matplotlib==3.10.8
numpy==2.4.1
opencv_python==4.13.0.90
Pillow==12.1.0
torch==2.5.1+cu121
tqdm==4.67.1
transformers==4.46.3
accelerate==1.12.0